+++ lib/events/ramp_closed.dart
@@ -1,5 +1,0 @@
-part of '../widget_event.dart';
-
-final class RampClosed extends WidgetEvent {
-  const RampClosed();
-}